Clean up BaGet's Startup
