+++ src/BaGet.Core/Configuration/MirrorOptions.cs
@@ -5,6 +5,12 @@ namespace BaGet.Core.Configuration
     public class MirrorOptions
     {
         /// <summary>
+        /// If true, packages that aren't found locally will be indexed
+        /// using the upstream source.
+        /// </summary>
+        public bool Enabled { get; set; }
+
+        /// <summary>
         /// The v3 index that will be mirrored.
         /// </summary>
         public Uri PackageSource { get; set; }
@@ -12,12 +18,6 @@ namespace BaGet.Core.Configuration
         /// <summary>
         /// The time before a download from the package source times out.
         /// </summary>
-        public int PackageDownloadTimeoutSeconds { get; set; }
-
-        /// <summary>
-        /// If true, packages that aren't found locally will be indexed
-        /// using the upstream source.
-        /// </summary>
-        public bool EnableReadThroughCaching { get; set; }
+        public int PackageDownloadTimeoutSeconds { get; set; } = 600;
     }
 }
