+++ src/BaGet/Extensions/ServiceCollectionExtensions.cs
@@ -1,18 +1,26 @@
 ﻿using System;
+using System.IO;
 using System.Net;
 using System.Net.Http;
+using BaGet.Azure.Configuration;
+using BaGet.Azure.Extensions;
+using BaGet.Azure.Search;
+using BaGet.Configurations;
 using BaGet.Core.Configuration;
 using BaGet.Core.Entities;
 using BaGet.Core.Services;
 using BaGet.Entities;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Cors.Infrastructure;
+using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
 namespace BaGet.Extensions
 {
-    // TODO: Considering moving this to BaGet.Core
     public static class ServiceCollectionExtensions
     {
         public static void AddBaGetContext(this IServiceCollection services)
@@ -56,6 +64,84 @@ namespace BaGet.Extensions
             });
         }
 
+        public static void ConfigureHttpServices(this IServiceCollection services)
+        {
+            services.AddMvc();
+            services.AddCors();
+            services.AddSingleton<IConfigureOptions<CorsOptions>, ConfigureCorsOptions>();
+
+            services.Configure<ForwardedHeadersOptions>(options =>
+            {
+                options.ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto;
+
+                // Do not restrict to local network/proxy
+                options.KnownNetworks.Clear();
+                options.KnownProxies.Clear();
+            });
+        }
+
+        public static void ConfigureStorageProviders(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.Configure<FileSystemStorageOptions>(configuration.GetSection(nameof(BaGetOptions.Storage)));
+
+            services.AddTransient<IPackageStorageService>(provider =>
+            {
+                var options = provider.GetRequiredService<IOptions<BaGetOptions>>().Value;
+
+                switch (options.Storage.Type)
+                {
+                    case StorageType.FileSystem:
+                        return provider.GetRequiredService<FilePackageStorageService>();
+
+                    case StorageType.AzureBlobStorage:
+                        return provider.GetRequiredService<BlobPackageStorageService>();
+
+                    default:
+                        throw new InvalidOperationException(
+                            $"Unsupported storage service: {options.Storage.Type}");
+                }
+            });
+
+            services.AddTransient(provider =>
+            {
+                var options = provider.GetRequiredService<IOptions<FileSystemStorageOptions>>().Value;
+                var path = string.IsNullOrEmpty(options.Path)
+                    ? Path.Combine(Directory.GetCurrentDirectory(), "Packages")
+                    : options.Path;
+
+                // Ensure the package storage directory exists
+                Directory.CreateDirectory(path);
+
+                return new FilePackageStorageService(path);
+            });
+
+            services.AddBlobPackageStorageService();
+        }
+
+        public static void ConfigureSearchProviders(this IServiceCollection services)
+        {
+            services.AddTransient<ISearchService>(provider =>
+            {
+                var options = provider.GetRequiredService<IOptions<BaGetOptions>>().Value;
+
+                switch (options.Search.Type)
+                {
+                    case SearchType.Database:
+                        return provider.GetRequiredService<DatabaseSearchService>();
+
+                    case SearchType.Azure:
+                        return provider.GetRequiredService<AzureSearchService>();
+
+                    default:
+                        throw new InvalidOperationException(
+                            $"Unsupported search service: {options.Search.Type}");
+                }
+            });
+
+            services.AddTransient<DatabaseSearchService>();
+            services.AddAzureSearch();
+        }
+
         /// <summary>
         /// Add the services that mirror an upstream package source.
         /// </summary>
@@ -66,7 +152,7 @@ namespace BaGet.Extensions
             {
                 var options = provider.GetRequiredService<IOptions<BaGetOptions>>().Value;
 
-                if (!options.Mirror.EnableReadThroughCaching)
+                if (!options.Mirror.Enabled)
                 {
                     return new FakeMirrorService();
                 }
@@ -93,6 +179,16 @@ namespace BaGet.Extensions
                 client.Timeout = TimeSpan.FromSeconds(options.Mirror.PackageDownloadTimeoutSeconds);
 
                 return client;
+            });
+        }
+
+        public static void ConfigureAuthenticationProviders(this IServiceCollection services)
+        {
+            services.AddSingleton<IAuthenticationService, ApiKeyAuthenticationService>(provider =>
+            {
+                var options = provider.GetRequiredService<IOptions<BaGetOptions>>().Value;
+
+                return new ApiKeyAuthenticationService(options.ApiKeyHash);
             });
         }
     }
