+++ src/BaGet/Startup.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.IO;
-using BaGet.Azure.Configuration;
 using BaGet.Azure.Extensions;
-using BaGet.Azure.Search;
+using BaGet.Configurations;
 using BaGet.Core.Configuration;
 using BaGet.Core.Entities;
 using BaGet.Core.Services;
@@ -10,130 +8,41 @@ using BaGet.Extensions;
 using BaGet.Web.Extensions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
 
 namespace BaGet
 {
     public class Startup
     {
-        private const string CorsPolicy = "AllowAll";
-
         public Startup(IConfiguration configuration)
         {
-            Configuration = configuration;
+            Configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
         }
 
         public IConfiguration Configuration { get; }
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc();
-            services.AddBaGetContext();
-
             services.Configure<BaGetOptions>(Configuration);
-            services.Configure<FileSystemStorageOptions>(Configuration.GetSection(nameof(BaGetOptions.Storage)));
+
+            services.AddBaGetContext();
+            services.ConfigureHttpServices();
             services.ConfigureAzure(Configuration);
-
-            services.AddCors(options =>
-            {
-                options.AddPolicy(
-                    CorsPolicy,
-                    builder => builder.AllowAnyOrigin()
-                        .AllowAnyMethod()
-                        .AllowAnyHeader());
-            });
-            
-            services.Configure<ForwardedHeadersOptions>(options =>
-            {
-                options.ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto;
-                // Do not restrict to local network/proxy
-                options.KnownNetworks.Clear();
-                options.KnownProxies.Clear();
-            });
-
-            services.AddSingleton<IAuthenticationService, ApiKeyAuthenticationService>(provider =>
-            {
-                var options = provider.GetRequiredService<IOptions<BaGetOptions>>().Value;
-
-                return new ApiKeyAuthenticationService(options.ApiKeyHash);
-            });
 
             services.AddTransient<IPackageService, PackageService>();
             services.AddTransient<IIndexingService, IndexingService>();
             services.AddMirrorServices();
 
-            ConfigureStorageProviders(services);
-            ConfigureSearchProviders(services);
-        }
-
-        private void ConfigureStorageProviders(IServiceCollection services)
-        {
-            services.AddTransient<IPackageStorageService>(provider =>
-            {
-                var options = provider.GetRequiredService<IOptions<BaGetOptions>>().Value;
-
-                switch (options.Storage.Type)
-                {
-                    case StorageType.FileSystem:
-                        return provider.GetRequiredService<FilePackageStorageService>();
-
-                    case StorageType.AzureBlobStorage:
-                        return provider.GetRequiredService<BlobPackageStorageService>();
-
-                    default:
-                        throw new InvalidOperationException(
-                            $"Unsupported storage service: {options.Storage.Type}");
-                }
-            });
-
-            services.AddTransient(provider =>
-            {
-                var options = provider.GetRequiredService<IOptions<FileSystemStorageOptions>>().Value;
-                var path = string.IsNullOrEmpty(options.Path)
-                    ? Path.Combine(Directory.GetCurrentDirectory(), "Packages")
-                    : options.Path;
-
-                // Ensure the package storage directory exists
-                Directory.CreateDirectory(path);
-
-                return new FilePackageStorageService(path);
-            });
-
-            services.AddBlobPackageStorageService();
-        }
-
-        private void ConfigureSearchProviders(IServiceCollection services)
-        {
-            services.AddTransient<ISearchService>(provider =>
-            {
-                var options = provider.GetRequiredService<IOptions<BaGetOptions>>().Value;
-
-                switch (options.Search.Type)
-                {
-                    case SearchType.Database:
-                        return provider.GetRequiredService<DatabaseSearchService>();
-
-                    case SearchType.Azure:
-                        return provider.GetRequiredService<AzureSearchService>();
-
-                    default:
-                        throw new InvalidOperationException(
-                            $"Unsupported search service: {options.Search.Type}");
-                }
-            });
-
-            services.AddTransient<DatabaseSearchService>();
-            services.AddAzureSearch();
+            services.ConfigureStorageProviders(Configuration);
+            services.ConfigureSearchProviders();
+            services.ConfigureAuthenticationProviders();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
         {
-            app.UseForwardedHeaders();
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
@@ -152,10 +61,11 @@ namespace BaGet
                 }
             }
 
+            app.UseForwardedHeaders();
             app.UseDefaultFiles();
             app.UseStaticFiles();
 
-            app.UseCors(CorsPolicy);
+            app.UseCors(ConfigureCorsOptions.CorsPolicy);
 
             app.UseMvc(routes =>
             {
