+++ src/BaGet/appsettings.json
@@ -16,9 +16,8 @@
   },
 
   "Mirror": {
-    "PackageSource": "https://api.nuget.org/v3-flatcontainer/",
-    "PackageDownloadTimeoutSeconds": 600,
-    "EnableReadThroughCaching":  false
+    "Enabled": false,
+    "PackageSource": "https://api.nuget.org/v3-flatcontainer/"
   },
 
   "Logging": {
